imxrt-flash: fix reading data from AHB
Calculate correct address and return correct number of bytes.

JIRA: NIL-570

+++ storage/imxrt-flash/fspi.h
@@ -71,7 +71,7 @@
 
 typedef struct _flexspi_t {
 	volatile uint32_t *base;
-	uint32_t *ahbAddr;
+	uint8_t *ahbAddr;
 	uint8_t instance;
 	uint8_t slPortMask;
 	size_t slFlashSz[4];
